接口代码更新尝试

+++ app/src/main/java/xyz/looorielovbb/playground/pojo/ListResp.kt
@@ -7,5 +7,5 @@ data class ListResp<T>(
     val over: Boolean,
     val pageCount: Int,
     val size: Int,
-    val total: Long,
+    val total: Int,
 )
